:recycle: 関数分割
