+++ src/main.c
@@ -12,7 +12,7 @@
 sig_atomic_t g_sig = 0;
 
 
-void free_tokens(t_token *token)
+void free_tokens(t_token *token)////////////////
 {
 	t_token *temp;
